// src/schemas/register.schema.ts
import * as z from "zod";

export const registerSchema = z.object({
  name: z.string().min(2, "Name must be at least 2 characters"),
  email: z.email("Invalid email address"),
  password: z.string().min(6, "Password must be at least 6 characters"),
  role: z.enum(["NUTRITIONIST", "STUDENT"]).default("NUTRITIONIST"),
});

export type RegisterFormData = z.infer<typeof registerSchema>;